@@ -1,7 +1,9 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
-import 'package:flutter_circular_chart/flutter_circular_chart.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:flutter_ui_app/global/responsive.dart';
+import 'package:syncfusion_flutter_charts/charts.dart';
 
 class Tasks extends StatelessWidget {
   @override
@@ -13,14 +15,14 @@ class Tasks extends StatelessWidget {
         height: responsive.inchR(30),
         width: double.infinity,
         child: Column(
-          children: <Widget>[TasksTitle(), TaskList()],
+          children: <Widget>[_TasksTitle(), _TaskList()],
         ),
       ),
     );
   }
 }
 
-class TasksTitle extends StatelessWidget {
+class _TasksTitle extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final responsive = Responsive.of(context);
@@ -79,19 +81,19 @@ class TasksTitle extends StatelessWidget {
   }
 }
 
-class TaskList extends StatelessWidget {
+class _TaskList extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Expanded(
         child: Row(
             mainAxisAlignment: MainAxisAlignment.spaceEvenly,
             children: <Widget>[
-          const Task(
+          const _Task(
               percent: 50,
               taskName: 'Investigar protocolos \nde internet',
               taskTile: 'Seguridad informatica',
               userImage: 'assets/canvas/images/img_person_1.jpg'),
-          const Task(
+          const _Task(
               percent: 20,
               taskName: 'Investigar arquitectura \nde computadoras',
               taskTile: 'Partes de computadoras',
@@ -100,13 +102,13 @@ class TaskList extends StatelessWidget {
   }
 }
 
-class Task extends StatelessWidget {
-  const Task(
-      {Key key,
-      @required this.percent,
-      @required this.taskName,
-      @required this.taskTile,
-      @required this.userImage})
+class _Task extends StatelessWidget {
+  const _Task(
+      {Key? key,
+      required this.percent,
+      required this.taskName,
+      required this.taskTile,
+      required this.userImage})
       : super(key: key);
 
   final double percent;
@@ -151,7 +153,7 @@ class Task extends StatelessWidget {
               ),
             ],
           ),
-          TaskUser(
+          _TaskUser(
             userImage: userImage,
           )
         ],
@@ -161,7 +163,7 @@ class Task extends StatelessWidget {
 }
 
 class TaskName extends StatelessWidget {
-  const TaskName({Key key, @required this.name}) : super(key: key);
+  const TaskName({Key? key, required this.name}) : super(key: key);
 
   final String name;
 
@@ -184,45 +186,46 @@ class TaskName extends StatelessWidget {
 }
 
 class TaskChart extends StatelessWidget {
-  const TaskChart({Key key, @required this.percent}) : super(key: key);
+  const TaskChart({Key? key, required this.percent}) : super(key: key);
 
   final double percent;
 
   @override
   Widget build(BuildContext context) {
-    final _chartKey = GlobalKey<AnimatedCircularChartState>();
-
     final responsive = Responsive.of(context);
-
-    return Container(
-      padding: const EdgeInsets.all(5),
-      width: responsive.inchR(10),
-      height: responsive.inchR(10),
-      child: AnimatedCircularChart(
-        key: _chartKey,
-        size: Size(responsive.inchR(8), responsive.inchR(8)),
-        startAngle: 240,
-        initialChartData: <CircularStackEntry>[
-          CircularStackEntry(
-            <CircularSegmentEntry>[
-              CircularSegmentEntry(
-                percent,
-                const Color(0xff367BE2),
+    return Padding(
+      padding: const EdgeInsets.only(left: 5),
+      child: Container(
+        padding: const EdgeInsets.all(12),
+        width: responsive.inchR(10),
+        height: responsive.inchR(10),
+        child: Stack(
+          children: [
+            Container(
+              height: double.infinity,
+              width: double.infinity,
+              child: CustomPaint(
+                  painter: _RadialProgress(
+                      percentage: percent,
+                      primaryColor: const Color(0xff367BE2),
+                      secondaryColor: const Color(0xffDCEBF5),
+                      primaryLine: 3)),
+            ),
+            Container(
+              height: double.infinity,
+              width: double.infinity,
+              child: Center(
+                child: Text(
+                  '$percent%',
+                  style: TextStyle(
+                    color: const Color(0xff367BE2),
+                    fontWeight: FontWeight.bold,
+                    fontSize: responsive.inchR(1.6),
+                  ),
+                ),
               ),
-              CircularSegmentEntry(
-                100 - percent,
-                const Color(0xffDCEBF5),
-              ),
-            ],
-          ),
-        ],
-        chartType: CircularChartType.Radial,
-        percentageValues: true,
-        holeLabel: '$percent%',
-        labelStyle: TextStyle(
-          color: const Color(0xff367BE2),
-          fontWeight: FontWeight.bold,
-          fontSize: responsive.inchR(1.6),
+            )
+          ],
         ),
       ),
     );
@@ -230,7 +233,7 @@ class TaskChart extends StatelessWidget {
 }
 
 class TaskTile extends StatelessWidget {
-  const TaskTile({Key key, @required this.tile}) : super(key: key);
+  const TaskTile({Key? key, required this.tile}) : super(key: key);
 
   final String tile;
 
@@ -267,8 +270,8 @@ class TaskTile extends StatelessWidget {
   }
 }
 
-class TaskUser extends StatelessWidget {
-  const TaskUser({Key key, this.userImage}) : super(key: key);
+class _TaskUser extends StatelessWidget {
+  const _TaskUser({Key? key, required this.userImage}) : super(key: key);
 
   final String userImage;
 
@@ -289,4 +292,49 @@ class TaskUser extends StatelessWidget {
       ),
     );
   }
+}
+
+class _RadialProgress extends CustomPainter {
+  _RadialProgress({
+    required this.percentage,
+    required this.primaryColor,
+    required this.secondaryColor,
+    required this.primaryLine,
+  });
+
+  final double percentage;
+  final Color primaryColor;
+  final Color secondaryColor;
+  final double primaryLine;
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final center = Offset(size.width * 0.5, size.height / 2);
+    final radio = min(size.width * 0.5, size.height * 0.5);
+
+    final paintArco = Paint()
+      ..strokeWidth = primaryLine
+      ..color = primaryColor
+      ..strokeCap = StrokeCap.round
+      ..style = PaintingStyle.stroke;
+
+    final paintArco2 = Paint()
+      ..strokeWidth = primaryLine
+      ..color = secondaryColor
+      ..strokeCap = StrokeCap.round
+      ..style = PaintingStyle.stroke;
+
+    final arcAngle = 2 * pi * (percentage / 100);
+    canvas.drawArc(Rect.fromCircle(center: center, radius: radio), -pi / 2,
+        arcAngle, false, paintArco);
+
+    final arcAngle2 =
+        -(2 * pi * ((93 - ((percentage == 100 ? 93 : percentage))) / 100));
+
+    canvas.drawArc(Rect.fromCircle(center: center, radius: radio), -pi / 1.75,
+        arcAngle2, false, paintArco2);
+  }
+
+  @override
+  bool shouldRepaint(CustomPainter oldDelegate) => true;
 }
