@@ -4,7 +4,7 @@ import 'package:flutter_ui_app/apps/canvas/pages/to_do_page/widgets/widgets.dart
 import 'package:flutter_ui_app/global/responsive.dart';
 
 class ActivityList extends StatelessWidget {
-  const ActivityList({Key key, @required this.activities}) : super(key: key);
+  const ActivityList({Key? key, required this.activities}) : super(key: key);
 
   final List<Activity> activities;
   @override
