@@ -1,9 +1,7 @@
-import 'package:flutter/widgets.dart';
-
 class ContactItem {
   ContactItem({
-    @required this.name,
-    @required this.email,
+    required this.name,
+    required this.email,
   });
 
   String name;
